gen-ai/tutorials/evaluation/getting_started/pyproject.toml
[project]
name = "getting-started-evaluation"
version = "0.0.0"
description = "Getting started with GenAI Evaluator SDK"
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.0,<2.0.0",
    "gllm-evals[deepeval,langchain,ragas]>=0.0.41",
]

[[tool.uv.index]]
name = "gen-ai-internal"
url = "https://glsdk.gdplabs.id/gen-ai-internal/simple/"

[tool.uv.sources]
gllm-evals = { index = "gen-ai-internal" }
